move shareQR function to QrCode file
We're going to use this function for the onboarding flow. And besides, we don't need it cluttering up ProfileSettings anyway. Probably best to extract it to a general location, and work with any input message.
