+++ www/js/control/ProfileSettings.jsx
@@ -21,6 +21,7 @@ import ControlSyncHelper, {ForceSyncRow, getHelperSyncSettings} from "./ControlS
 import ControlCollectionHelper, {getHelperCollectionSettings, getState, isMediumAccuracy, helperToggleLowAccuracy, forceTransition} from "./ControlCollectionHelper";
 import { resetDataAndRefresh } from "../config/dynamicConfig";
 import { AppContext } from "../App";
+import { shareQR } from "../components/QrCode";
 
 //any pure functions can go outside
 const ProfileSettings = () => {
@@ -261,35 +262,6 @@ const ProfileSettings = () => {
         refreshCollectSettings();
     }
 
-    const shareQR = function() {
-       /*code adapted from demo of react-qr-code*/
-        const svg = document.querySelector(".qr-code");
-        const svgData = new XMLSerializer().serializeToString(svg);
-        const img = new Image();
-
-        img.onload = () => {
-            const canvas = document.createElement("canvas");
-            const ctx = canvas.getContext("2d");
-            canvas.width = img.width;
-            canvas.height = img.height;
-            ctx.drawImage(img, 0, 0);
-            const pngFile = canvas.toDataURL("image/png");
-                
-            var prepopulateQRMessage = {}; 
-            prepopulateQRMessage.files = [pngFile];
-            prepopulateQRMessage.url = authSettings.opcode;
-            prepopulateQRMessage.message = authSettings.opcode; //text saved to files with image!
-                
-            window.plugins.socialsharing.shareWithOptions(prepopulateQRMessage, function(result) {
-                console.log("Share completed? " + result.completed); // On Android apps mostly return false even while it's true
-                console.log("Shared to app: " + result.app); // On Android result.app is currently empty. On iOS it's empty when sharing is cancelled (result.completed=false)
-            }, function(msg) {
-                console.log("Sharing failed with message: " + msg);
-            });
-        }
-        img.src =  `data:image/svg+xml;base64,${btoa(svgData)}`;
-    }
-
     const viewQRCode = function(e) {
         setOpCodeVis(true);
     }
@@ -442,7 +414,7 @@ const ProfileSettings = () => {
             <ActionMenu vis={forceStateVis} setVis={setForceStateVis} title={"Force State"} actionSet={stateActions} onAction={onSelectState} onExit={() => {}}></ActionMenu>
 
             {/* opcode viewing popup */}
-            <PopOpCode visibilityValue = {opCodeVis} setVis = {setOpCodeVis} tokenURL = {"emission://login_token?token="+authSettings.opcode} action={shareQR}></PopOpCode>
+            <PopOpCode visibilityValue = {opCodeVis} setVis = {setOpCodeVis} tokenURL = {"emission://login_token?token="+authSettings.opcode} action={() => shareQR(authSettings.opcode)}></PopOpCode>
 
             {/* {view privacy} */}
             <PrivacyPolicyModal privacyVis={privacyVis} setPrivacyVis={setPrivacyVis}></PrivacyPolicyModal>
